requirements.txt
streamlit==1.24.1
ultralytics==8.0.30
torch>=1.12.0
numpy>=1.21.0
Pillow>=8.0.0
